+++ 07.bowling_object/shot.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+class Shot
+  attr_reader :mark
+
+  def initialize(mark)
+    @mark = mark
+  end
+
+  def score
+    return 10 if mark == 'X'
+
+    mark.to_i
+  end
+end
